Crate gu_ethkey

Ethereum keys management supporting keystores in formats used by geth, parity and pyethereum.

Features

keys generation

keys serialization/deserialization

keystore password change

signing/verification

encryption/decryption

Example

extern crate gu_ethkey;
use gu_ethkey::prelude::*;

fn main() {
    let key = SafeEthKey::load_or_generate("/tmp/path/to/keystore", &"passwd".into())
        .expect("should load or generate new eth key");

    println!("{:?}", key.address())
}

Modules

prelude
A "prelude" for users of the gu-ethkey crate.

Structs

Error
The Error type.

SafeEthKey
An Ethereum KeyPair wrapper with Store.

Enums

ErrorKind
The kind of an error.

Constants

KEY_ITERATIONS
HMAC fn iteration count; compromise between security and performance

Traits

EthKey
Provides basic EC operations on curve Secp256k1.

EthKeyStore
Provides basic serde for Ethereum KeyPair.

ResultExt
Additional methods for Result, for easy interaction with this crate.

Type Definitions

Result
Convenient wrapper around std::Result.